--- src/main/resources/logback.xml
@@ -1,34 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<configuration>
-    <appender name="LOCAL" class="ch.qos.logback.core.ConsoleAppender">
-        <encoder>
-            <!-- <pattern>%-5level [%thread] %d{HH:mm:ss.SSS} %logger{15} - %msg%n%rEx</pattern> -->
-            <!-- <pattern>%date{yyyy-MM-dd} %X{akkaTimestamp} %-5level[%thread] %logger{1} - %msg%n</pattern> -->
-            <!-- <pattern>%-5level %logger{36}: %msg%n</pattern> -->
-            <!-- <pattern>%d{HH:mm:ss.SSS} [%thread] %-5level %class{36}.%M %L - %msg%n</pattern> -->
-            <pattern>%d{HH:mm:ss.SSS} [%thread] %-5level %logger{36} - %msg%n</pattern>
-        </encoder>
-    </appender>
-
-    <logger name="akka" level="DEBUG"/>
-
-    <root level="DEBUG">
-        <appender-ref ref="LOCAL"/>
-    </root>
-
-    <logger name="com.zaxxer" level="INFO"/>
-
-    <logger name="com.pinkstack.realestate.NepClient" level="DEBUG" />
-
-
-
-    <!--
-    <logger name="akka" level="${AKKA_LOG_LEVEL:-INFO}" />
-    <logger name="oshi.util" level="INFO" />
-
-    <root level="${ROOT_LOG_LEVEL:-INFO}">
-      <appender-ref ref="${DEPLOYMENT:-GCP}" />
-      <appender-ref ref="BUGSNAG"/>
-    </root>
-    -->
-</configuration>
