--- worksheets/ranges.sc
@@ -1,1 +1,0 @@
-Range(1, 10).toList
